+++ database/update_task_status.sql
@@ -1,7 +1,5 @@
--- Функция для обновления статуса задачи
-
 CREATE OR REPLACE FUNCTION update_task_status(
-    p_task_id INTEGER,
+    p_task_id BIGINT,
     p_status VARCHAR,
     p_progress INTEGER DEFAULT NULL,
     p_progress_message TEXT DEFAULT NULL,
